UserAddress - Added attribute 'show_to_user' for simulate deletion

+++ app/Http/Controllers/API/v1/User/UserAddress/UserAddressDeleteController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\API\v1\User\UserAddress;
 
 use App\Http\Controllers\Controller;
+use App\Http\Resources\API\v1\User\UserResource;
 use App\Models\Address;
 use App\Models\User;
 
@@ -10,8 +11,16 @@ class UserAddressDeleteController extends Controller
 {
     public function __invoke(User $user, Address $address)
     {
-        $address->delete();
+        //имитирую удаление для пользователя, если есть заказы привязанные к адресу
 
-        return 'OK';
+        $addressOrders = $address->orders;
+        $ordersCount = $addressOrders->count();
+
+        if ($ordersCount == 0)
+            $address->delete();
+        else
+            $address->update(['show_to_user' => false]);
+
+        return new UserResource($user);
     }
 }
